糸巻で倒された敵の遅延演出処理を入れた、糸巻で倒された敵のスコア計算を入れた。スコア計算システムをインスタンスに変更

+++ Source/Game/ScissorsGame/ScissorsPlayerStateDerived.cpp
@@ -347,7 +347,7 @@ void ScissorsPlayerChargeDashState::Execute(float deltaTime)
     }
 
     int arrowCount = _countof(player->arrowComponents);
-    int segmentCount = player->dashPoints.size() - 1;
+    int segmentCount = static_cast<int>(player->dashPoints.size()) - 1;
 
     int count = std::min<int>(arrowCount, segmentCount);
